third_party/blink/web_tests/svg/dom/SVGNumberList-basics-expected.txt

ABC

This is a test of the simple SVGNumberList API parts.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS text1.rotate.baseVal != text1.rotate.animVal is true
PASS text1.rotate.baseVal.numberOfItems is 3
PASS text1.rotate.animVal.numberOfItems is 3


Test uncommon arguments for initialize()
PASS text1.rotate.baseVal.initialize(30) threw exception TypeError: Failed to execute 'initialize' on 'SVGNumberList': parameter 1 is not of type 'SVGNumber'..
PASS text1.rotate.baseVal.initialize('aString') threw exception TypeError: Failed to execute 'initialize' on 'SVGNumberList': parameter 1 is not of type 'SVGNumber'..
PASS text1.rotate.baseVal.initialize(text1) threw exception TypeError: Failed to execute 'initialize' on 'SVGNumberList': parameter 1 is not of type 'SVGNumber'..
PASS text1.rotate.baseVal.initialize(null) threw exception TypeError: Failed to execute 'initialize' on 'SVGNumberList': parameter 1 is not of type 'SVGNumber'..


Test uncommon arguments for getItem()
PASS text1.rotate.baseVal.getItem(30) threw exception IndexSizeError: Failed to execute 'getItem' on 'SVGNumberList': The index provided (30) is greater than the maximum bound (3)..
PASS text1.rotate.baseVal.getItem('aString') is text1.rotate.baseVal.getItem(0)
PASS text1.rotate.baseVal.getItem(text1) is text1.rotate.baseVal.getItem(0)
PASS text1.rotate.baseVal.getItem(null) is text1.rotate.baseVal.getItem(0)


Test uncommon arguments for insertItemBefore()
PASS text1.rotate.baseVal.insertItemBefore(30) threw exception TypeError: Failed to execute 'insertItemBefore' on 'SVGNumberList': 2 arguments required, but only 1 present..
PASS text1.rotate.baseVal.insertItemBefore('aString') threw exception TypeError: Failed to execute 'insertItemBefore' on 'SVGNumberList': 2 arguments required, but only 1 present..
PASS text1.rotate.baseVal.insertItemBefore(text1) threw exception TypeError: Failed to execute 'insertItemBefore' on 'SVGNumberList': 2 arguments required, but only 1 present..
PASS text1.rotate.baseVal.insertItemBefore(null) threw exception TypeError: Failed to execute 'insertItemBefore' on 'SVGNumberList': 2 arguments required, but only 1 present..
PASS text1.rotate.baseVal.insertItemBefore(text1.rotate.baseVal.getItem(1), 'aString') is text1.rotate.baseVal.getItem(0)
PASS text1.rotate.baseVal.numberOfItems is 4
PASS text1.rotate.baseVal.getItem(0).value is 180
PASS text1.rotate.baseVal.getItem(1).value is 90
PASS text1.rotate.baseVal.getItem(2).value is 180
PASS text1.rotate.baseVal.getItem(3).value is 270
PASS text1.getAttribute('rotate') is "180 90 180 270"
PASS text1.rotate.baseVal.insertItemBefore(text1.rotate.baseVal.getItem(1), text1) is text1.rotate.baseVal.getItem(0)
PASS text1.rotate.baseVal.numberOfItems is 5
PASS text1.rotate.baseVal.getItem(0).value is 90
PASS text1.rotate.baseVal.getItem(1).value is 180
PASS text1.rotate.baseVal.getItem(2).value is 90
PASS text1.rotate.baseVal.getItem(3).value is 180
PASS text1.rotate.baseVal.getItem(4).value is 270
PASS text1.getAttribute('rotate') is "90 180 90 180 270"
PASS text1.rotate.baseVal.insertItemBefore(text1.rotate.baseVal.getItem(1), null) is text1.rotate.baseVal.getItem(0)
PASS text1.rotate.baseVal.numberOfItems is 6
PASS text1.rotate.baseVal.getItem(0).value is 180
PASS text1.rotate.baseVal.getItem(1).value is 90
PASS text1.rotate.baseVal.getItem(2).value is 180
PASS text1.rotate.baseVal.getItem(3).value is 90
PASS text1.rotate.baseVal.getItem(4).value is 180
PASS text1.rotate.baseVal.getItem(5).value is 270
PASS text1.getAttribute('rotate') is "180 90 180 90 180 270"
PASS text1.rotate.baseVal.insertItemBefore(text1.rotate.baseVal.getItem(1), 0) is text1.rotate.baseVal.getItem(0)
PASS text1.rotate.baseVal.numberOfItems is 7
PASS text1.rotate.baseVal.getItem(0).value is 90
PASS text1.rotate.baseVal.getItem(1).value is 180
PASS text1.rotate.baseVal.getItem(2).value is 90
PASS text1.rotate.baseVal.getItem(3).value is 180
PASS text1.rotate.baseVal.getItem(4).value is 90
PASS text1.rotate.baseVal.getItem(5).value is 180
PASS text1.rotate.baseVal.getItem(6).value is 270
PASS text1.getAttribute('rotate') is "90 180 90 180 90 180 270"
PASS text1.rotate.baseVal.insertItemBefore(30, 0) threw exception TypeError: Failed to execute 'insertItemBefore' on 'SVGNumberList': parameter 1 is not of type 'SVGNumber'..
PASS text1.rotate.baseVal.insertItemBefore('aString', 0) threw exception TypeError: Failed to execute 'insertItemBefore' on 'SVGNumberList': parameter 1 is not of type 'SVGNumber'..
PASS text1.rotate.baseVal.insertItemBefore(text1, 0) threw exception TypeError: Failed to execute 'insertItemBefore' on 'SVGNumberList': parameter 1 is not of type 'SVGNumber'..
PASS text1.rotate.baseVal.insertItemBefore(null, 0) threw exception TypeError: Failed to execute 'insertItemBefore' on 'SVGNumberList': parameter 1 is not of type 'SVGNumber'..


Set rotate='1 2 3 4' for text1
PASS text1.setAttribute('rotate', '1 2 3 4') is undefined.
PASS text1.rotate.baseVal.numberOfItems is 4
PASS text1.rotate.baseVal.getItem(0).value is 1
PASS text1.rotate.baseVal.getItem(1).value is 2
PASS text1.rotate.baseVal.getItem(2).value is 3
PASS text1.rotate.baseVal.getItem(3).value is 4
PASS text1.getAttribute('rotate') is "1 2 3 4"


Test uncommon arguments for replaceItem()
PASS text1.rotate.baseVal.replaceItem(30) threw exception TypeError: Failed to execute 'replaceItem' on 'SVGNumberList': 2 arguments required, but only 1 present..
PASS text1.rotate.baseVal.replaceItem('aString') threw exception TypeError: Failed to execute 'replaceItem' on 'SVGNumberList': 2 arguments required, but only 1 present..
PASS text1.rotate.baseVal.replaceItem(text1) threw exception TypeError: Failed to execute 'replaceItem' on 'SVGNumberList': 2 arguments required, but only 1 present..
PASS text1.rotate.baseVal.replaceItem(null) threw exception TypeError: Failed to execute 'replaceItem' on 'SVGNumberList': 2 arguments required, but only 1 present..
PASS text1.rotate.baseVal.replaceItem(30, 0) threw exception TypeError: Failed to execute 'replaceItem' on 'SVGNumberList': parameter 1 is not of type 'SVGNumber'..
PASS text1.rotate.baseVal.replaceItem('aString', 0) threw exception TypeError: Failed to execute 'replaceItem' on 'SVGNumberList': parameter 1 is not of type 'SVGNumber'..
PASS text1.rotate.baseVal.replaceItem(text1, 0) threw exception TypeError: Failed to execute 'replaceItem' on 'SVGNumberList': parameter 1 is not of type 'SVGNumber'..
PASS text1.rotate.baseVal.replaceItem(null, 0) threw exception TypeError: Failed to execute 'replaceItem' on 'SVGNumberList': parameter 1 is not of type 'SVGNumber'..


Test uncommon values for indexed setter
PASS text1.rotate.baseVal[0] = 30 threw exception TypeError: Failed to set an indexed property [0] on 'SVGNumberList': parameter 2 is not of type 'SVGNumber'..
PASS text1.rotate.baseVal[0] = 'aString' threw exception TypeError: Failed to set an indexed property [0] on 'SVGNumberList': parameter 2 is not of type 'SVGNumber'..
PASS text1.rotate.baseVal[0] = text1 threw exception TypeError: Failed to set an indexed property [0] on 'SVGNumberList': parameter 2 is not of type 'SVGNumber'..
PASS text1.rotate.baseVal[0] = null threw exception TypeError: Failed to set an indexed property [0] on 'SVGNumberList': parameter 2 is not of type 'SVGNumber'..
PASS text1.rotate.baseVal.replaceItem(text1.rotate.baseVal.getItem(0), 0) is text1.rotate.baseVal.getItem(0)
PASS text1.rotate.baseVal.numberOfItems is 4
PASS text1.rotate.baseVal.getItem(0).value is 1
PASS text1.rotate.baseVal.getItem(1).value is 2
PASS text1.rotate.baseVal.getItem(2).value is 3
PASS text1.rotate.baseVal.getItem(3).value is 4
PASS text1.getAttribute('rotate') is "1 2 3 4"
PASS text1.rotate.baseVal.replaceItem(text1.rotate.baseVal.getItem(0), 'aString') is text1.rotate.baseVal.getItem(0)
PASS text1.rotate.baseVal.numberOfItems is 4
PASS text1.rotate.baseVal.getItem(0).value is 1
PASS text1.rotate.baseVal.getItem(1).value is 2
PASS text1.rotate.baseVal.getItem(2).value is 3
PASS text1.rotate.baseVal.getItem(3).value is 4
PASS text1.getAttribute('rotate') is "1 2 3 4"
PASS text1.rotate.baseVal.replaceItem(text1.rotate.baseVal.getItem(0), text1) is text1.rotate.baseVal.getItem(0)
PASS text1.rotate.baseVal.numberOfItems is 4
PASS text1.getAttribute('rotate') is "1 2 3 4"
PASS text1.rotate.baseVal.replaceItem(text1.rotate.baseVal.getItem(0), null) is text1.rotate.baseVal.getItem(0)
PASS text1.rotate.baseVal.numberOfItems is 4
PASS text1.getAttribute('rotate') is "1 2 3 4"


Set rotate='1 2 3 4' for text1
PASS text1.setAttribute('rotate', '1 2 3 4') is undefined.


Test uncommon arguents for removeItem()
PASS text1.rotate.baseVal.removeItem(30) threw exception IndexSizeError: Failed to execute 'removeItem' on 'SVGNumberList': The index provided (30) is greater than the maximum bound (4)..
PASS text1.rotate.baseVal.removeItem('aString').value is 1
PASS text1.rotate.baseVal.numberOfItems is 3
PASS text1.rotate.baseVal.getItem(0).value is 2
PASS text1.rotate.baseVal.getItem(1).value is 3
PASS text1.rotate.baseVal.getItem(2).value is 4
PASS text1.getAttribute('rotate') is "2 3 4"
PASS text1.rotate.baseVal.removeItem(text1).value is 2
PASS text1.rotate.baseVal.numberOfItems is 2
PASS text1.rotate.baseVal.getItem(0).value is 3
PASS text1.rotate.baseVal.getItem(1).value is 4
PASS text1.getAttribute('rotate') is "3 4"
PASS text1.rotate.baseVal.removeItem(null).value is 3
PASS text1.rotate.baseVal.numberOfItems is 1
PASS text1.rotate.baseVal.getItem(0).value is 4
PASS text1.getAttribute('rotate') is "4"


Test uncommon arguments for appendItem()
PASS text1.rotate.baseVal.appendItem(30) threw exception TypeError: Failed to execute 'appendItem' on 'SVGNumberList': parameter 1 is not of type 'SVGNumber'..
PASS text1.rotate.baseVal.appendItem('aString') threw exception TypeError: Failed to execute 'appendItem' on 'SVGNumberList': parameter 1 is not of type 'SVGNumber'..
PASS text1.rotate.baseVal.appendItem(text1) threw exception TypeError: Failed to execute 'appendItem' on 'SVGNumberList': parameter 1 is not of type 'SVGNumber'..
PASS text1.rotate.baseVal.appendItem(null) threw exception TypeError: Failed to execute 'appendItem' on 'SVGNumberList': parameter 1 is not of type 'SVGNumber'..


Testing animVal clear throws
PASS text1.rotate.animVal.clear() threw exception NoModificationAllowedError: Failed to execute 'clear' on 'SVGNumberList': This object is read-only..
PASS text1.rotate.baseVal.clear() is undefined.
PASS successfullyParsed is true

TEST COMPLETE
